+++ source/optix/amadeus_repg.cu
@@ -104,7 +104,7 @@ extern "C" __global__ void __raygen__()
 
 	// TODO: seed generatoin method
 	rp.seed = make_float3(
-		sin(idx.x - idx.y),
+		sinf(idx.x - idx.y),
 		parameters.samples,
 		parameters.time
 	);
